feat: goodbye old ui
